refactor(Core UI): :recycle: Use the button as a dateTimeRange label
