+++ package.json
@@ -57,7 +57,8 @@
     "@commitlint/config-conventional": "^17.7.0",
     "@cypress/webpack-preprocessor": "^6.0.0",
     "@testing-library/dom": "^9.3.3",
-    "@testing-library/jest-dom": "^5.17.0",
+    "@types/jest": "^29.5.5",
+    "@testing-library/jest-dom": "^6.1.3",
     "@testing-library/react": "^14.0.0",
     "@testing-library/user-event": "^14.5.1",
     "@typescript-eslint/eslint-plugin": "^6.7.4",
