+++ src/App.css
@@ -153,7 +153,6 @@ html {
 }
 
 /* refresh metrics button */
-
 .button-toggle-dropdown-loading button {
   background-color: var(--pf-v5-global--danger-color--100) !important;
 }
