+++ src/__tests__/App.spec.tsx
@@ -1,3 +1,4 @@
+import '@testing-library/jest-dom';
 import { render, screen, waitFor } from '@testing-library/react';
 import { Server } from 'miragejs';
